+++ CHANGELOG.md
@@ -1,9 +1,19 @@
 # bot-sdk 变更记录
 
+* Request新增获取apiAccessToken方法
+* Bot 新增添加权限回调方法
+* 修复权限bug
+
+***
+* 优化日志功能
+* 优化数据统计功能
+
+***
 * 增加Display.PushStack指令
 * 增加expectResponse
 * 增加Record.RecordSpeech指令
 * 增加Permission.AskForPermissionsConsent指令
+
 *** 
 * 增加列表标签Tag
 * 增加BodyTemplate6